Added Value Buffers
This is to make sure when you need to buffer a single value it is being done uniformly.
There is also a large value buffer which is designed to handle any large values stored on their own as there is a small optimization that can be made on large objects.