Whooo boy (New EntityBuffer)
Lots of little changes with buffers and registries getting new functions like rename which was available in a few and then the new EntityBuffer is here.

This new buffer works in the capacity of the old entity buffer and the old link buffer but in a way that will hopefully make finding links for systems faster.